% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cc_names.R
\name{camel_caser}
\alias{camel_caser}
\title{camelcaser}
\usage{
camel_caser(names)
}
\arguments{
\item{names}{data frame column names}
}
\description{
Function to convert data frame names to camelCase
}
\examples{
camel_caser(names(iris))

}
