% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pc_names.R
\name{paschal_caser}
\alias{paschal_caser}
\title{PaschalCaser}
\usage{
paschal_caser(names)
}
\arguments{
\item{names}{data frame column names}
}
\description{
Function to convert data frame names to PaschalCase
}
\examples{
paschal_caser(names(iris))

}
