% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rf_names.R
\name{reformat_names}
\alias{reformat_names}
\title{reformat_names}
\usage{
reformat_names(names, format = "snake")
}
\arguments{
\item{names}{data frame column names}

\item{format}{name format}
}
\description{
Function to convert data frame names to a standardized format
}
\examples{
reformat_names(names(iris), 'snake')

}
