% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc_names.R
\name{snake_caser}
\alias{snake_caser}
\title{snakecaser}
\usage{
snake_caser(names)
}
\arguments{
\item{names}{data frame column names}
}
\description{
Function to convert data frame names to snake_case
}
\examples{
snake_caser(names(iris))

}
