% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tc_names.R
\name{titlecaser}
\alias{titlecaser}
\alias{title_caser}
\title{Convert data frame column names to Title Case}
\usage{
title_caser(names)
}
\arguments{
\item{names}{data frame column names}
}
\description{
Function to convert data frame names to title names for export
}
\examples{
title_caser(names(iris))

}
