% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cc_names.R
\name{to_camel_case}
\alias{to_camel_case}
\title{Convert data frame column names to camelCase}
\usage{
to_camel_case(names)
camel_caser(names)
}
\arguments{
\item{names}{data frame column names}
}
\description{
Function to convert data frame names to camelCase
}
\examples{
to_camel_case(names(iris))
camel_caser(names(iris))
}
