% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc_names.R
\name{to_snake_case}
\alias{to_snake_case}
\title{Convert data frame column names to snake_case}
\usage{
to_snake_case(names)
snake_caser(names)
}
\arguments{
\item{names}{data frame column names}
}
\description{
Function to convert data frame names to snake_case
}
\examples{
to_snake_case(names(iris))
snake_caser(names(iris))
}
