% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pc_names.R
\name{paschal_caser}
\alias{paschal_caser}
\alias{to_pascal_case}
\title{Convert data frame column names to PaschalCase}
\usage{
to_pascal_case(names)
}
\arguments{
\item{names}{data frame column names}
}
\description{
Function to convert data frame names to PaschalCase
}
\details{
Convert data frame column names to PascalCase
}
\examples{
to_pascal_case(names(iris))


}
