% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_df_names.R
\name{print_df_names}
\alias{print_df_names}
\title{Print data frame variable names}
\usage{
print_df_names(df, regex_pattern, comma)
}
\arguments{
\item{df}{name of data frame to print column names}

\item{regex_pattern}{optional argument to use a regular 
expression to filter column names}

\item{comma}{optional argument to seperate column names
with a comma}
}
\description{
This function prints data frame column names in the console in an easy to copy and paste format
}
\examples{
print_df_names('iris')
}
\keyword{column}
\keyword{names}
